+++ src/main/webapp/components/sort-order/index.js
@@ -1,0 +1,2 @@
+import SortOrderGQL, { SortOrder } from './sort-order'
+export { SortOrderGQL as default, SortOrder }
