+++ src/main/webapp/query-builder/filter-inputs/date-filter.tsx
@@ -6,7 +6,7 @@ const {
   TimeRangeDuring,
   TimeRangeRelative,
   validate,
-} = require('../../time-range')
+} = require('../../components/time-range')
 const {
   uglyMap,
   parseRelative,
